LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/columnGap-expected.txt


PASS column-gap responsive to style changes
FAIL column-gap clamped to 0px assert_equals: expected "0px" but got "20px"
FAIL column-gap responsive to inherited changes assert_equals: expected "80px" but got "normal"
